Add RTI_NC_LICENSE_ACCEPTED environment variable to buildfarm builds.
This config value relies on features in https://github.com/ros-infrastructure/ros_buildfarm/pull/554 for functionality but I do not think anything will break if it's added now.

This value is only needed on amd64 as there is no RTI package for arm64. The 'yes' is quoted to avoid it being interpreted by pyyaml as a boolean which would cause it to manifest as `True` rather than `yes` in the generated Dockerfile.