Switch rolling sync threshold from count to required packages.
For most ROS distributions we want to periodically ratchet up the
package count for sync thresholds. However Rolling is meant to be
somewhat volatile as large changes are made.

A large enough number of ancillary packages have been introduced to
Rolling that the "base" numeric package count is no longer sufficient to
guarantee the presence of ROS core packages. So let's instead use the
required packages sync criteria and require the presence of ros_desktop
which will in turn require the ROS core and a relatively healthy ROS
system to be available.