Enable shared ccache on Rolling debug job.
Investigating whether enabling ccache with the current build
configuration improves build performance.
On ci.ros2.org the pool of agents is even more semi-permanent than on
build.ros2.org.

This may only be a temporary palliative as we will eventually transition
to using the generic agents for CI jobs using the job-weight attributes.