Rename Ubuntu arm64 build file.
Keeping the exact target name in the build file name is going to cause
unecessary churn in commit history as the build file is renamed each
time the platform rolls.

Since the build file key in the index and the file contents both include
target information the filename does not need to as well.