Re-add libssl-dev to explicit packages for cross-vendor CI
The upstream job of the cross-vendor CI jobs is the 'nightly-release'
job, which includes the upstream ROS 1 packages explicitly so that the
bridge can be built and tested. In the cross-vendor jobs, the ROS 1
packages are not installed because the bridge is not built or tested.
One of the ROS 1 dependencies in the nightly-release jobs is pulling in
libssl-dev as a dependency. Fast-DDS is detecting libssl-dev and using
it to build security extensions automatically.

Because libssl-dev isn't a formal dependency of Fast-DDS and the ROS 1
packages aren't listed in the cross-vendor jobs, libssl-dev isn't
getting installed and any attempts to link against Fast-DDS fail due to
the absence.

An alternative to this change would be to explicitly disable security in
Fast-DDS so that libssl-dev is not used even when it is present.

Signed-off-by: Scott K Logan <logans@cottsay.net>