Specifically install tinyxml2 in cross-vendor CI
Fast-DDS doesn't install its manifest, so the dependency from Fast-DDS
on tinyxml2 is missed in these downstream jobs.

The dependency is listed in the colcon index, but the colcon index
doesn't list the package type so we can't know that the dependencies
could be consumed like rosdep keys. Fast-DDS isn't an ament package and
therefore isn't in the ament index either.

Signed-off-by: Scott K Logan <logans@cottsay.net>