Add failing packages to build blacklists on Rolling and Humble.
These packages are failing to build on Rolling Ubuntu amd64
- [swri_console_util](https://build.ros2.org/job/Rbin_uJ64__swri_console_util__ubuntu_jammy_amd64__binary)
- [swri_geometry_util](https://build.ros2.org/job/Rbin_uJ64__swri_geometry_util__ubuntu_jammy_amd64__binary)
- [swri_prefix_tools](https://build.ros2.org/job/Rbin_uJ64__swri_prefix_tools__ubuntu_jammy_amd64__binary)
- [swri_roscpp](https://build.ros2.org/job/Rbin_uJ64__swri_roscpp__ubuntu_jammy_amd64__binary)

These packages are failing to build on Rolling RHEL 8 x86_64
- [swri_console_util](https://build.ros2.org/job/Rbin_rhel_el864__swri_console_util__rhel_8_x86_64__binary)
- [swri_geometry_util](https://build.ros2.org/job/Rbin_rhel_el864__swri_geometry_util__rhel_8_x86_64__binary)
- [swri_prefix_tools](https://build.ros2.org/job/Rbin_rhel_el864__swri_prefix_tools__rhel_8_x86_64__binary)
- [swri_roscpp](https://build.ros2.org/job/Rbin_rhel_el864__swri_roscpp__rhel_8_x86_64__binary)

And on Humble Ubuntu amd64
- [swri_console_util](https://build.ros2.org/job/Hbin_uJ64__swri_console_util__ubuntu_jammy_amd64__binary)
- [swri_geometry_util](https://build.ros2.org/job/Hbin_uJ64__swri_geometry_util__ubuntu_jammy_amd64__binary)
- [swri_prefix_tools](https://build.ros2.org/job/Hbin_uJ64__swri_prefix_tools__ubuntu_jammy_amd64__binary)
- [swri_roscpp](https://build.ros2.org/job/Hbin_uJ64__swri_roscpp__ubuntu_jammy_amd64__binary)

And Humble RHEL 8
- [swri_console_util](https://build.ros2.org/job/Hbin_rhel_el864__swri_console_util__rhel_8_x86_64__binary)
- [swri_geometry_util](https://build.ros2.org/job/Hbin_rhel_el864__swri_geometry_util__rhel_8_x86_64__binary)
- [swri_prefix_tools](https://build.ros2.org/job/Hbin_rhel_el864__swri_prefix_tools__rhel_8_x86_64__binary)
- [swri_roscpp](https://build.ros2.org/job/Hbin_rhel_el864__swri_roscpp__rhel_8_x86_64__binary)

Normally for packages which are failing to build globally the first
thing I would do is remove the package from the release list in the
distribution files. That approach has the advantage of allowing a simple
re-release to trigger a new attempt to build the blocked packages.
However, when using that approach downstream jobs are not detected and
blocked along with the blacklisted job which means that packages which
depend on "dropped" releases will just start failing with an error
installing the dropped package.

I've come around to the idea that it's better for the build farm overall
if packages are blocked via build configuration unless the entire
release is broken and should be removed.