Add ouster_msgs to excluded packages on Rolling and Humble.
This is a followup to #236 and #237.
Although this package does not directly or indirectly depend on
libtins-dev, when bloom fails to resolve rosdep keys for the sibling
package ros2_ouster it drops all release metadata for the platform in
that repository regardless of individual package.

As a result the sourcepkg job for RHEL fails since the expected tags are
missing in the bloom repository.

It is possible that a future enhancement to bloom could process sibling
packages individually and allow ouster_msgs to continue building in this
situation but for now the only way to get ouster_msgs building on RHEL
would be to split it out into its own repository.